+++ assets/model_monitoring/components/tests/e2e/test_genai_preprocessor_e2e.py
@@ -16,7 +16,7 @@ from tests.e2e.utils.constants import (
 
 def _submit_genai_preprocessor_job(
     submit_pipeline_job, ml_client: MLClient, get_component, experiment_name,
-    input_data, start_time, end_time
+    input_data, start_time, end_time, expect_failure: bool = False
 ):
     genai_preprocessor_component = get_component(COMPONENT_NAME_GENAI_PREPROCESSOR)
 
@@ -48,7 +48,7 @@ def _submit_genai_preprocessor_job(
     )
 
     pipeline_job = submit_pipeline_job(
-        pipeline_job, experiment_name
+        pipeline_job, experiment_name, expect_failure
     )
 
     # Wait until the job completes