+++ scripts/azureml-assets/CHANGELOG.md
@@ -3,6 +3,26 @@
 
 ### 🐛 Bugs Fixed
 
+## 1.16.50 (2024-05-22)
+### 🐛 Bugs Fixed
+- [#2956](https://github.com/Azure/azureml-assets/pull/2956) Update embeddings asset task type
+
+## 1.16.49 (2024-05-17)
+### 🐛 Bugs Fixed
+- [#2935](https://github.com/Azure/azureml-assets/pull/2935) Update the manifest file to include additional configs
+
+## 1.16.48 (2024-05-17)
+### 🐛 Bugs Fixed
+- [#2934](https://github.com/Azure/azureml-assets/pull/2934) Update asset validation check for EvaluationResult
+
+## 1.16.47 (2024-05-17)
+### 🐛 Bugs Fixed
+- [#2933](https://github.com/Azure/azureml-assets/pull/2933) skip hiddenlayerscanned tags till most models are scanned
+
+## 1.16.46 (2024-05-15)
+### 🐛 Bugs Fixed
+- [#2907](https://github.com/Azure/azureml-assets/pull/2907) add hiddenlayerscanned tags valdn
+
 ## 1.16.45 (2024-04-24)
 ### 🐛 Bugs Fixed
 - [#2773](https://github.com/Azure/azureml-assets/pull/2773) Add h100 skus to supported inference skus list